Create unified timeline/surveycard components
