+++ packages/website/src/routes/ReefRoutes/Reef/Surveys/DeleteButton/index.tsx
@@ -13,10 +13,10 @@ import DeleteOutlineIcon from "@material-ui/icons/DeleteOutline";
 import CloseIcon from "@material-ui/icons/Close";
 import moment from "moment";
 
-import surveyServices from "../../../../services/surveyServices";
-import { userInfoSelector } from "../../../../store/User/userSlice";
-import { surveysRequest } from "../../../../store/Survey/surveyListSlice";
-import DeleteDialog, { Action } from "../../../../common/Dialog";
+import surveyServices from "../../../../../services/surveyServices";
+import { userInfoSelector } from "../../../../../store/User/userSlice";
+import { surveysRequest } from "../../../../../store/Survey/surveyListSlice";
+import DeleteDialog, { Action } from "../../../../../common/Dialog";
 
 const DeleteButton = ({
   reefId,
