--- packages/website/src/routes/ReefRoutes/Reef/Surveys/types.ts
@@ -1,8 +1,0 @@
-import { SurveyMedia } from "../../../../store/Survey/types";
-
-export interface TimelineProps {
-  reefId: number;
-  isAdmin: boolean;
-  observation: SurveyMedia["observations"] | "any";
-  point: number;
-}
